Move mappr_server to mappr_server_kbase
Also server_kbase now behaives differently:
Initialy, if not already existing, a room called ubiquitous is saved
into the database. It is meant to receive all viewpoints as containing
room, since room creation is not yet supported by the rviz plugin.

If a viewpoint already exists adding is aborted. This is done to enforce
using the add service call for adding only and the update service call
for changing existing viewpoints.